🚚 reorganize testing utils
