--- packages/atom.io/test/-util.ts
@@ -1,7 +1,0 @@
-export const silence: Pick<Console, `error` | `info` | `warn`> = {
-  error: () => null,
-  warn: () => null,
-  info: () => null,
-}
-
-export const stdout = (..._: any[]): void => undefined
