+++ packages/atom.io/test/__util__/stdout.ts
@@ -1,0 +1,1 @@
+export const stdout = (..._: any[]): void => undefined
