+++ packages/atom.io/test/store-observation.test.ts
@@ -1,18 +1,8 @@
 import { vitest } from "vitest"
 
-import * as UTIL from "./-util"
-import {
-  __INTERNAL__,
-  atom,
-  getState,
-  runTransaction,
-  selector,
-  setLogLevel,
-  setState,
-  subscribe,
-  transaction,
-} from "../src"
-import { redo, timeline, undo } from "../src/timeline"
+import * as UTIL from "./__util__"
+import { __INTERNAL__, atom, selector, setLogLevel, transaction } from "../src"
+import { timeline } from "../src/timeline"
 
 const LOG_LEVELS = [null, `error`, `warn`, `info`] as const
 const CHOOSE = 0
