+++ packages/atom.io/test/transaction-internal.test.ts
@@ -2,7 +2,7 @@
 
 import { vitest } from "vitest"
 
-import * as UTIL from "./-util"
+import * as UTIL from "./__util__"
 import {
   __INTERNAL__,
   atom,
